Keep runs view dispatch out of CLI root
